NLP Prompt Data Validation Framework/requirement.txt
python-dotenv
google-generativeai
reportlab
pandas